requirements.txt
imap
PyMuPDF
openai==0.28.0
mysql-connector-python